MyNumber.doubleValue returns a 'double' instead of 'java.lang.Double'

+++ org.example.xtext.modeldsl.parent/org.example.xtext.modeldsl/xtend-gen/org/example/xtext/runtime/MyNumber.java
@@ -8,7 +8,7 @@ public class MyNumber {
     this.value = value;
   }
   
-  public Double doubleValue() {
-    return this.value;
+  public double doubleValue() {
+    return this.value.doubleValue();
   }
 }
